Improve result output of YacqRun, add macro member to show member list
Implement sequence output in YacqRun, dump some elements of the result if it is
IEnumerable (and is not String).
Add macro members (Object.? and Static<Object>.?) to show suitable member and
methods (in the type and symbol table).
Separate type matching algorithm in SymbolEntry.Match to SymbolEntry (TypeMatch
methods) and fix some bugs in the code: not to match all Static<> symbols if
LeftType is Object, not to match all global symbols in macro symbols context.