Support to define member, function, and method symbols
Extend symbol-handling symbols ((def), (def!), and (undef) functions and
methods) to support to define member, function, and method symbols, like
(def [Int32].(method) ...) . The definition of the symbol also accepts
three-argument ambiguous lambda for define not a literal but a
SymbolDefinition delegate.
Fix a bug in SymbolEntry class, only uses TargetMask value in DispatchType
property in Equals and GetHashCode methods (ignore DispatchTypes.Literal value
in comparing).
Refactor the code in NumberExpression class.
Add new function in YacqExpression.List extension method, if "head" argument is
null, or not supplied, then returns all the elements of the list, or null if it
is not a list.