Add reader context handlings
Modify the rules in built-in grammars to handle reader the context stack.
Add transiting expression rules for the pattern grammar.
Fix the transiting expression rule to the standard grammar in the alternative
grammar (re-implement as an extended term).
Improve error outputs of YacqRepl to use information from the reader context.